src/inference/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.41.0      # INT4/INT8 quantization for smaller GPUs
timm>=0.9.0               # Vision encoder (SigLIP)
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
numpy>=1.24.0
httpx>=0.27.0             # async HTTP client for testing
python-multipart>=0.0.9   # file uploads (camera frames)
